Implemented the Join Command
ClientCommand has a bug on msg_error() where the inheeritance of the function
is not working... Must fix it

The Client and the Player are independant of each other. The Player is not
necesarilly a Client, this is because Players exist even if a Client doesn't

I didn't understood this at first but now I see this. To implement something
that goes more according to this new perception, intead of inheriting the
Client from the Player. A player should have a `Client * client_;` field, which
might or not be `NULL`